dvb: Try pre-DVBv5 card query first, use as fallback.
It seems that some drivers (as seen in #2727)
do not implement delivery-system enumeration but otherwise
can still be used via DVBv5 API.
Implement behaviour similar to popular w_scan:
First, query with the old ioctl which can only return one DELSYS,
and if the DVBv5 ioctl fails, fallback to that.
If the DVBv5 ioctl works, discard the result from the first query.